Rolling

+++ Assets/Scripts/DamagePlayer.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace sg {
     public class DamagePlayer : MonoBehaviour {
-        private int damage = 25;
+        private float damage = 25;
         private void OnTriggerEnter(Collider other) {
             PlayerStats playerStats = other.GetComponent<PlayerStats>();
             if (playerStats != null) {
