+++ Assets/Scripts/Managers/CharacterStats.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 namespace sg {
     public class CharacterStats : MonoBehaviour {
         public int healthLevel = 10;
-        public int maxHealth;
-        public int currentHealth;
+        public float maxHealth;
+        public float currentHealth;
 
         public int staminaLevel = 10;
-        public int maxStamina;
-        public int currentStamina;
+        public float maxStamina;
+        public float currentStamina;
 
         public bool isDead;
     }
